Added jne statements to procs row1_cmp-row3_cmp to fix broken payouts
If the numbers procedures row1_cmp, row2_cmp, or row3_cmp are comparing are not equal, it jumps to the next row comparison or returns from the function after row3. This fixed a bug where it was paying out the user even when they did not get 3 of the same number in a row.